+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Clickable bot command suggestions, email addresses and hashtags in messages
 - Clickable links in messages on the map
 - Tray icon now shows a blue circle over the logo in case of unread messsages (only linux + win)
+- reload profile image on `DC_EVENT_SELFAVATAR_CHANGED` in settings
 
 ### Changed
 - Bring back the back and close button in dialogs, move actions that need an ok/cancel button into it's own dialogs